backend/Message_Passing_Module/agent_assigner/requirements.txt
functions-framework==3.*
google-cloud-pubsub>=2.22.0
google-cloud-firestore>=2.16.1
